+++ fesm5/core.js
@@ -1,5 +1,5 @@
 /**
- * @license Angular v8.0.0-beta.1+109.sha-a227c52
+ * @license Angular v8.0.0-beta.2+3.sha-35e45dc
  * (c) 2010-2019 Google LLC. https://angular.io/
  * License: MIT
  */
@@ -4324,14 +4324,6 @@ function getCheckNoChangesMode() {
 function setCheckNoChangesMode(mode) {
     checkNoChangesMode = mode;
 }
-/** Whether or not this is the first time the current view has been processed. */
-var firstTemplatePass = true;
-function getFirstTemplatePass() {
-    return firstTemplatePass;
-}
-function setFirstTemplatePass(value) {
-    firstTemplatePass = value;
-}
 /**
  * The root index from which pure function instructions should calculate their binding
  * indices. In component views, this is TView.bindingStartIndex. In a host binding
@@ -4373,7 +4365,6 @@ function enterView(newView, hostTNode) {
     var oldView = lView;
     if (newView) {
         var tView = newView[TVIEW];
-        firstTemplatePass = tView.firstTemplatePass;
         bindingRootIndex = tView.bindingStartIndex;
     }
     previousOrParentTNode = hostTNode;
@@ -7090,13 +7081,13 @@ function getLViewChild(lView) {
  * @param view The view to be destroyed.
  */
 function destroyLView(view) {
-    var renderer = view[RENDERER];
-    if (isProceduralRenderer(renderer) && renderer.destroyNode) {
-        walkTNodeTree(view, 2 /* Destroy */, renderer, null);
+    if (!(view[FLAGS] & 128 /* Destroyed */)) {
+        var renderer = view[RENDERER];
+        if (isProceduralRenderer(renderer) && renderer.destroyNode) {
+            walkTNodeTree(view, 2 /* Destroy */, renderer, null);
+        }
+        destroyViewTree(view);
     }
-    destroyViewTree(view);
-    // Sets the destroyed flag
-    view[FLAGS] |= 128 /* Destroyed */;
 }
 /**
  * Determines which LViewOrLContainer to jump to when traversing back up the
@@ -7133,6 +7124,12 @@ function getParentState(state, rootView) {
 function cleanUpView(viewOrContainer) {
     if (viewOrContainer.length >= HEADER_OFFSET) {
         var view = viewOrContainer;
+        // Mark the LView as destroyed *before* executing the onDestroy hooks. An onDestroy hook
+        // runs arbitrary user code, which could include its own `viewRef.destroy()` (or similar). If
+        // We don't flag the view as destroyed before the hooks, this could lead to an infinite loop.
+        // This also aligns with the ViewEngine behavior. It also means that the onDestroy hook is
+        // really more of an "afterDestroy" hook if you think about it.
+        view[FLAGS] |= 128 /* Destroyed */;
         executeOnDestroys(view);
         removeListeners(view);
         var hostTNode = view[HOST_NODE];
@@ -9067,7 +9064,6 @@ function refreshDescendantViews(lView) {
     var tView = lView[TVIEW];
     // This needs to be set before children are processed to support recursive components
     tView.firstTemplatePass = false;
-    setFirstTemplatePass(false);
     // Resetting the bindingIndex of the current LView as the next steps may trigger change detection.
     lView[BINDING_INDEX] = tView.bindingStartIndex;
     // If this is a creation pass, we should not call lifecycle hooks or evaluate bindings.
@@ -9270,7 +9266,6 @@ function renderEmbeddedTemplate(viewToRender, tView, context) {
             // off firstTemplatePass. If we don't set it here, instances will perform directive
             // matching, etc again and again.
             viewToRender[TVIEW].firstTemplatePass = false;
-            setFirstTemplatePass(false);
             refreshDescendantViews(viewToRender);
         }
         finally {
@@ -9505,7 +9500,7 @@ function createDirectivesAndLocals(tView, lView, localRefs, localRefExtractor) {
     if (!getBindingsEnabled())
         return;
     var previousOrParentTNode = getPreviousOrParentTNode();
-    if (getFirstTemplatePass()) {
+    if (tView.firstTemplatePass) {
         ngDevMode && ngDevMode.firstTemplatePass++;
         resolveDirectives(tView, lView, findDirectiveMatches(tView, lView, previousOrParentTNode), previousOrParentTNode, localRefs || null);
     }
@@ -10402,7 +10397,7 @@ function instantiateRootComponent(tView, viewData, def) {
  */
 function resolveDirectives(tView, viewData, directives, tNode, localRefs) {
     // Please make sure to have explicit type for `exportsMap`. Inferred type triggers bug in tsickle.
-    ngDevMode && assertEqual(getFirstTemplatePass(), true, 'should run on first template pass only');
+    ngDevMode && assertEqual(tView.firstTemplatePass, true, 'should run on first template pass only');
     var exportsMap = localRefs ? { '': -1 } : null;
     if (directives) {
         initNodeFlags(tNode, tView.data.length, directives.length);
@@ -10437,7 +10432,7 @@ function resolveDirectives(tView, viewData, directives, tNode, localRefs) {
 function instantiateAllDirectives(tView, lView, tNode) {
     var start = tNode.directiveStart;
     var end = tNode.directiveEnd;
-    if (!getFirstTemplatePass() && start < end) {
+    if (!tView.firstTemplatePass && start < end) {
         getOrCreateNodeInjectorForNode(tNode, lView);
     }
     for (var i = start; i < end; i++) {
@@ -10453,7 +10448,7 @@ function invokeDirectivesHostBindings(tView, viewData, tNode) {
     var start = tNode.directiveStart;
     var end = tNode.directiveEnd;
     var expando = tView.expandoInstructions;
-    var firstTemplatePass = getFirstTemplatePass();
+    var firstTemplatePass = tView.firstTemplatePass;
     for (var i = start; i < end; i++) {
         var def = tView.data[i];
         var directive = viewData[i];
@@ -10495,7 +10490,7 @@ function generateExpandoInstructionBlock(tView, tNode, directiveCount) {
 */
 function prefillHostVars(tView, lView, totalHostVars) {
     ngDevMode &&
-        assertEqual(getFirstTemplatePass(), true, 'Should only be called in first template pass.');
+        assertEqual(tView.firstTemplatePass, true, 'Should only be called in first template pass.');
     for (var i = 0; i < totalHostVars; i++) {
         lView.push(NO_CHANGE);
         tView.blueprint.push(NO_CHANGE);
@@ -10537,7 +10532,7 @@ function postProcessBaseDirective(lView, previousOrParentTNode, directive, def) 
 * If a component is matched (at most one), it is returned in first position in the array.
 */
 function findDirectiveMatches(tView, viewData, tNode) {
-    ngDevMode && assertEqual(getFirstTemplatePass(), true, 'should run on first template pass only');
+    ngDevMode && assertEqual(tView.firstTemplatePass, true, 'should run on first template pass only');
     var registry = tView.directiveRegistry;
     var matches = null;
     if (registry) {
@@ -10563,9 +10558,9 @@ function findDirectiveMatches(tView, viewData, tNode) {
 }
 /** Stores index of component's host element so it will be queued for view refresh during CD. */
 function queueComponentIndexForCheck(previousOrParentTNode) {
-    ngDevMode &&
-        assertEqual(getFirstTemplatePass(), true, 'Should only be called in first template pass.');
     var tView = getLView()[TVIEW];
+    ngDevMode &&
+        assertEqual(tView.firstTemplatePass, true, 'Should only be called in first template pass.');
     (tView.components || (tView.components = [])).push(previousOrParentTNode.index);
 }
 /**
@@ -10574,7 +10569,7 @@ function queueComponentIndexForCheck(previousOrParentTNode) {
 */
 function queueHostBindingForCheck(tView, def, hostVars) {
     ngDevMode &&
-        assertEqual(getFirstTemplatePass(), true, 'Should only be called in first template pass.');
+        assertEqual(tView.firstTemplatePass, true, 'Should only be called in first template pass.');
     var expando = tView.expandoInstructions;
     var length = expando.length;
     // Check whether a given `hostBindings` function already exists in expandoInstructions,
@@ -10625,7 +10620,6 @@ function saveNameToExportMap(index, def, exportsMap) {
  * @param index the initial index
  */
 function initNodeFlags(tNode, index, numberOfDirectives) {
-    ngDevMode && assertEqual(getFirstTemplatePass(), true, 'expected firstTemplatePass to be true');
     var flags = tNode.flags;
     ngDevMode && assertEqual(flags === 0 || flags === 1 /* isComponent */, true, 'expected node flags to not be initialized');
     ngDevMode && assertNotEqual(numberOfDirectives, tNode.directiveEnd - tNode.directiveStart, 'Reached the max number of directives');
@@ -10653,7 +10647,7 @@ function addComponentLogic(lView, previousOrParentTNode, def) {
     // so this is a regular element, wrap it with the component view
     componentView[HOST] = lView[previousOrParentTNode.index];
     lView[previousOrParentTNode.index] = componentView;
-    if (getFirstTemplatePass()) {
+    if (lView[TVIEW].firstTemplatePass) {
         queueComponentIndexForCheck(previousOrParentTNode);
     }
 }
@@ -10773,7 +10767,7 @@ function template(index, templateFn, consts, vars, tagName, attrs, localRefs, lo
     var tView = lView[TVIEW];
     // TODO: consider a separate node type for templates
     var tNode = containerInternal(index, tagName || null, attrs || null);
-    if (getFirstTemplatePass()) {
+    if (tView.firstTemplatePass) {
         tNode.tViews = createTView(-1, templateFn, consts, vars, tView.directiveRegistry, tView.pipeRegistry, null);
     }
     createDirectivesAndLocals(tView, lView, localRefs, localRefExtractor);
@@ -10798,7 +10792,10 @@ function template(index, templateFn, consts, vars, tagName, attrs, localRefs, lo
  */
 function container(index) {
     var tNode = containerInternal(index, null, null);
-    getFirstTemplatePass() && (tNode.tViews = []);
+    var lView = getLView();
+    if (lView[TVIEW].firstTemplatePass) {
+        tNode.tViews = [];
+    }
     setIsParent(false);
 }
 function containerInternal(index, tagName, attrs) {
@@ -11171,11 +11168,10 @@ function projection(nodeIndex, selectorIndex, attrs) {
  */
 function addToViewTree(lView, adjustedHostIndex, state) {
     var tView = lView[TVIEW];
-    var firstTemplatePass = getFirstTemplatePass();
     if (lView[TAIL]) {
         lView[TAIL][NEXT] = state;
     }
-    else if (firstTemplatePass) {
+    else if (tView.firstTemplatePass) {
         tView.childIndex = adjustedHostIndex;
     }
     lView[TAIL] = state;
@@ -11395,10 +11391,10 @@ function bind(value) {
  * @param count Amount of vars to be allocated
  */
 function allocHostVars(count) {
-    if (!getFirstTemplatePass())
-        return;
     var lView = getLView();
     var tView = lView[TVIEW];
+    if (!tView.firstTemplatePass)
+        return;
     queueHostBindingForCheck(tView, getCurrentDirectiveDef(), count);
     prefillHostVars(tView, lView, count);
 }
@@ -13391,7 +13387,7 @@ var Version = /** @class */ (function () {
 /**
  * @publicApi
  */
-var VERSION = new Version('8.0.0-beta.1+109.sha-a227c52');
+var VERSION = new Version('8.0.0-beta.2+3.sha-35e45dc');
 
 /**
  * @license
@@ -19204,7 +19200,7 @@ read) {
     var tView = lView[TVIEW];
     var contentQuery = query(predicate, descend, read);
     (lView[CONTENT_QUERIES] || (lView[CONTENT_QUERIES] = [])).push(contentQuery);
-    if (getFirstTemplatePass()) {
+    if (tView.firstTemplatePass) {
         var tViewContentQueries = tView.contentQueries || (tView.contentQueries = []);
         var lastSavedDirectiveIndex = tView.contentQueries.length ? tView.contentQueries[tView.contentQueries.length - 1] : -1;
         if (directiveIndex !== lastSavedDirectiveIndex) {
